added lifetime for radon and fixed bug in convert to allow it lifetimes as floats instead of integers
Former-commit-id: 83984621656bad2617e8d3438f1377678c66d128

+++ acrg_time/convert.py
@@ -172,4 +172,4 @@ def julian2time(dates):
         
 def convert_to_hours(time):
     hours_per_unit = {"H": 1, "D": 24, "W": 168, "M": 732, "Y":8760}
-    return int(time[:-1]) * hours_per_unit[time[-1]]
+    return float(time[:-1]) * hours_per_unit[time[-1]]
